Fix warnings

+++ rrcounter/rrcounter.hh
@@ -38,7 +38,7 @@ class RoundRobinCounter {
         }
       }
       this->summation = 0;
-      for (int i = 0; i < _size; i++) {
+      for (size_t i = 0; i < _size; i++) {
         this->summation += this->slots[i];
       }
       this->last_index = current_index;
